+++ substratevm/src/com.oracle.svm.graal/src/com/oracle/svm/graal/stubs/AArch64StubForeignCallsFeature.java
@@ -32,7 +32,6 @@ import org.graalvm.nativeimage.Platforms;
 
 import com.oracle.svm.shared.feature.AutomaticallyRegisteredFeature;
 import com.oracle.svm.shared.singletons.traits.BuiltinTraits.BuildtimeAccessOnly;
-import com.oracle.svm.shared.singletons.traits.BuiltinTraits.Disallowed;
 import com.oracle.svm.shared.singletons.traits.BuiltinTraits.NoLayeredCallbacks;
 import com.oracle.svm.shared.singletons.traits.SingletonTraits;
 
@@ -72,7 +71,7 @@ import jdk.graal.compiler.replacements.nodes.VectorizedMismatchNode;
 
 @AutomaticallyRegisteredFeature
 @Platforms(AARCH64.class)
-@SingletonTraits(access = BuildtimeAccessOnly.class, layeredCallbacks = NoLayeredCallbacks.class, other = Disallowed.class)
+@SingletonTraits(access = BuildtimeAccessOnly.class, layeredCallbacks = NoLayeredCallbacks.class)
 public class AArch64StubForeignCallsFeature extends StubForeignCallsFeatureBase {
 
     public AArch64StubForeignCallsFeature() {